@@ -1,29 +1,25 @@
-﻿using System.Collections.Generic;
+﻿namespace ConversionExtensionsGenerator;
 
-namespace ConversionExtensionsGenerator
+public class BaseResult
 {
-    public class BaseResult
+    public List<GenerationLogItem> Errors { get; } = new();
+}
+
+public enum LogLevel
+{
+    Error,
+    Warning
+}
+
+public class GenerationLogItem
+{
+    public GenerationLogItem(LogLevel logLevel, string message)
     {
-        public List<GenerationLogItem> Errors { get; set; } = new List<GenerationLogItem>();
+        Message = message;
+        LogLevel = logLevel;
     }
 
-    public enum LogLevel
-    {
-        Error,
-        Warning
-    }
+    public LogLevel LogLevel { get; }
 
-    public class GenerationLogItem
-    {
-
-        public GenerationLogItem(LogLevel logLevel, string message)
-        {
-            Message = message;
-            LogLevel = logLevel;
-        }
-
-        public LogLevel LogLevel { get; }
-
-        public string Message { get; }
-    }
+    public string Message { get; }
 }
