@@ -1,125 +1,121 @@
-﻿using System;
-using System.Collections.Generic;
+﻿namespace ConversionExtensionsGenerator.Tests.Models;
 
-namespace ConversionExtensionsGenerator.Tests.Models
+public partial class PlainEntityModel
 {
-    public partial class PlainEntityModel
-    {
-        public int IntProperty { get; set; }
+    public int IntProperty { get; set; }
 
-        public long LongProperty { get; set; }
+    public long LongProperty { get; set; }
 
-        public string StringProperty { get; set; }
+    public string StringProperty { get; set; }
 
-        public string String2Property { get; set; }
+    public string String2Property { get; set; }
 
-        public virtual PlainEntity2Model PlainEntityClass2Property { get; set; }
+    public virtual PlainEntity2Model PlainEntityClass2Property { get; set; }
 
-        public virtual ICollection<PlainEntity2Model> PlainEntitiesCollectionProperty { get; set; }
+    public virtual ICollection<PlainEntity2Model> PlainEntitiesCollectionProperty { get; set; }
 
-        public int IntField;
+    public int IntField;
 
-        public long LongField;
+    public long LongField;
 
-        public string StringField;
+    public string StringField;
 
-        public string StringField2;
+    public string StringField2;
 
-        public PlainEntity2Model PlainEntityClass2Field;
+    public PlainEntity2Model PlainEntityClass2Field;
 
-        public ICollection<PlainEntity2Model> PlainEntitiesCollectionField;
-    }
+    public ICollection<PlainEntity2Model> PlainEntitiesCollectionField;
+}
 
-    public partial class PlainEntity2Model
-    {
-        public int IntProperty { get; set; }
+public partial class PlainEntity2Model
+{
+    public int IntProperty { get; set; }
 
-        public Guid GuidProperty { get; set; }
+    public Guid GuidProperty { get; set; }
 
-        public bool BoolProperty { get; set; }
+    public bool BoolProperty { get; set; }
 
-        public long LongProperty { get; set; }
+    public long LongProperty { get; set; }
 
-        public string StringProperty { get; set; }
+    public string StringProperty { get; set; }
 
-        public string StringProperty2 { get; set; }
+    public string StringProperty2 { get; set; }
 
-        public IEnumerable<PlainEntity3Model> PlainEntitiesCollectionProperty { get; set; }
+    public IEnumerable<PlainEntity3Model> PlainEntitiesCollectionProperty { get; set; }
 
-        public int IntField;
+    public int IntField;
 
-        public Guid GuidField;
+    public Guid GuidField;
 
-        public bool BoolField;
+    public bool BoolField;
 
-        public long LongField;
+    public long LongField;
 
-        public string StringField;
+    public string StringField;
 
-        public string StringField2;
+    public string StringField2;
 
-        public IEnumerable<PlainEntity3Model> PlainEntitiesCollectionField;
-    }
+    public IEnumerable<PlainEntity3Model> PlainEntitiesCollectionField;
+}
 
-    public partial class PlainEntity3Model
-    {
-        public Guid GuidProperty { get; set; }
+public partial class PlainEntity3Model
+{
+    public Guid GuidProperty { get; set; }
 
-        public bool BoolProperty { get; set; }
+    public bool BoolProperty { get; set; }
 
-        public long LongProperty { get; set; }
+    public long LongProperty { get; set; }
 
-        public string StringProperty { get; set; }
+    public string StringProperty { get; set; }
 
-        public TestEnumModel TestEnumProperty { get; set; }
+    public TestEnumModel TestEnumProperty { get; set; }
 
-        public TestEnumModel TestEnum3Property { get; set; }
+    public TestEnumModel TestEnum3Property { get; set; }
 
-        public Guid GuidField;
+    public Guid GuidField;
 
-        public bool BoolField;
+    public bool BoolField;
 
-        public long LongField;
+    public long LongField;
 
-        public string StringField;
+    public string StringField;
 
-        public TestEnumModel TestEnumField;
+    public TestEnumModel TestEnumField;
 
-        public TestEnumModel TestEnum3Field;
-    }
+    public TestEnumModel TestEnum3Field;
+}
 
-    public partial class EntitySomeModel
-    {
-        public Guid GuidProperty { get; set; }
+public partial class EntitySomeModel
+{
+    public Guid GuidProperty { get; set; }
 
-        public bool BoolProperty { get; set; }
+    public bool BoolProperty { get; set; }
 
-        public long LongProperty { get; set; }
+    public long LongProperty { get; set; }
 
-        public string StringProperty { get; set; }
+    public string StringProperty { get; set; }
 
-        public TestEnumModel TestEnumProperty { get; set; }
+    public TestEnumModel TestEnumProperty { get; set; }
 
-        public TestEnumModel TestEnumProperty2 { get; set; }
+    public TestEnumModel TestEnumProperty2 { get; set; }
 
-        public Guid GuidField;
+    public Guid GuidField;
 
-        public bool BoolField;
+    public bool BoolField;
 
-        public long LongField;
+    public long LongField;
 
-        public string StringField;
+    public string StringField;
 
-        public TestEnumModel TestEnumField;
+    public TestEnumModel TestEnumField;
 
-        public TestEnumModel TestEnumField2;
-    }
+    public TestEnumModel TestEnumField2;
+}
 
-    public enum TestEnumModel
-    {
-        One,
-        Two,
-        Three,
-        Four
-    }
+public enum TestEnumModel
+{
+    One,
+    Two,
+    Three,
+    Four
 }
